[F] S610-022 Use top as a default for value of a location.
Use top as a default value of a location instead of bot. Top as
a default value is safer than bot.

To be able to write something to a location representing the content
of an array, write bot to it when creating such location in malloc
if the identical location isn't already present in the abstract
state (otherwise, we are in a loop, the location represents multiple
allocsites and has to retain original values).

Context:

In bufferoverrun, there is a single location representing all elements of
an array. Since this location represents multiple values, it is always
weakly updated. Before this commit, the initial value of this location was
bottom (that is, the first update of this location in a procedure was
"strong").

TODO: Note that there are three kinds of locations representing multiple
values - a location representing all elements of an array and a location
containing a starfield. Currently, both are treated uniformly and the latter
ones are thus never updated with anything different from top. It would probably
make sense to have bottom as a default value for locations containign a
starfield.

Note that given the fact that we have a single location representing all
array elements, we are not able to capture the common pattern when an
element of an array is updated and read and there is not any update of
any other element of the array in between. For example:

   Arr (I) := 1;
   ... -- no update of Arr
   pragma Assert (Arr (I) = 2);

To be able to do that, we could use a recency abstraction. To represent
elements of an array, two locations and one value would be used.
First location representing values of all elements of the array, the second
location representing the value of the latest update of the array, and
the identification of the lastly updated element.

backport of commits 807582, eaf869, 7b640b, and 0352a271b.

807582: Use top as a default for value of a location.
Define Val.top. Except for traces, Val.top is defined the same way as
Val.unknown_from.

Note that the value Val.Itv.top which was used to represent top value is
top value only for non-pointers and non-arrays. Replace the use of
Val.Itv.top by Val.top in places where the value can be a pointer or an
array. This fixes problems when we were getting bottom values for
dereferences of unknown pointers and accesses to indexes of unknown
arrays.

Use Val.top as a value of a location which is not present in an abstract
state being joined. That is, we no longer assume that an abstract state
contains complete information (a location which can be updated in a
concrete execution doesn't have to be in an abstract state).
This fixes problems in particular with conditional assignments of input
variables of record types or global variables.

7b640b: Use top instead of bot if location is not in abstract state.
If a location is not in an abstract state, its value should be top.
In particular, after this change a weak update of a location which is
not in an abstract state will result in top instead of the value with
which the location is updated. Next, fix the use of Val.bot in widening of
abstract states.

This change is supposed to complete the change in the commit
807582d.

0352a27: T702-045 Fix updates of a location representing all elements of an array.
Write bot as a value of fresh allocsite.