[I] T221-007 Prune also stack variables.
backport of c920afa9:

Before this commit, when doing prune concerning a stack variable,
only a heap variable being an alias of this stack variable was pruned
and the stack variable was kept as it was. This is problem if the
stack variable is further used in the RHS of an assignment because
an unpruned value is assigned to the LHS location. This happens
for example when pruning Ada range checks after the change on
SB15-011. After this change, the RHS of an assignment is first assigned
to a temporary stack variable, this variable is pruned by the constraint
of the range check and then it is assigned to the LHS variable.