[I] T203-037 Fix producing bottom values when pruning.
backport of 67a2183:

When bufferoverrun analysis detects that a constraint implied by
a prune operation corresponding to a branching condition is not
satisfiable, the operation has to result in a bottom state so that
bufferoverrun checker can detect that the condition is always false.
In order to detect that pruning operation results in a bottom state,
bufferoverrun analysis checks whether pruning operation produces
an empty (bottom) value.

Before this fix, a prune operation could produce values which were
not considered as empty even if they represented result of pruning
with an unsatisfiable constraint.

Fix the problem by detecting whether a prune operation produced
an empty (bottom) interval or an empty (bottom) arrayblock and if
so return an empty (bottom) value instead of just returning
a value with an empty interval or an empty arrayblock (because
such value can be considered as non-empty).