[F] Fix aliasing of return value + use top as its initial value.
T214-011, T219-031.

Backport of the commits 1c9bf14 and 3493c52:

1c9bf14: T214-011 Fix aliasing of return value
Bufferoverrun computes information whether a return value of a function is
a must alias of some variable. However, before this fix it could happen
that bufferoverrun infered that a return value is a must alias of a
variable while it is only its may alias. This is because of the following:
(1) The default value of an alias of a return value of a function is
bottom.
(2) When analyzing an assignment to a variable representing a return value,
we didn't always update the information about an alias of the return value.
For example, we didn't do so if there was no aliasing information about
the RHS of the assignment.

That is, if there are two conditional assignments to a variable
representing a return value of a function and only one updates aliasing
information of return value, this aliasing information was joined with
bottom resulting in an incorrect must alias.

To fix that, always update the information about aliasing of a return
value when analyzing an assignment to return value. Use top if no alias
is found.

-----

3493c52: T219-031 Use top as initial value for alias of return value.
Use top instead of bottom as an initial value for an alias of a return
value. Using bottom can lead to incorrectly conclude that a variable is a
must alias of a return value of a function even though it is only a may
alias.

For example:

      function Test (B : Boolean; I : Int_Ptr) return Int_Ptr is
      begin
         if not B then
            raise Not_Valid_Error;
         end if;
         return I;  --  return value is only a may alias of I
      end Test;