[I] T507-015 More precise handling of symbolic intervals.
backport of 7013ab9:

Handle symbolic intervals in more precise way. In particular (1) improve
the detection of the fact that a symbolic value is precise, (2) improve
the conversion of (symbolic) intervals to boolean values, and (3)
improve the equality of (symbolic) intervals.

(1) is important for the detection of the fact that reachability
information is complete. Incomplete reachability information means that a
check cannot be propagated to a precondition and may lead to missing TP
messages.

The detection is improved by improving the function Bounds.are_similar
which is used in Val.is_precise. A symbolic value is (likely) precise if
its bounds are similar. Before the change, two bounds were considered
similar if they contained the same symbols. We now also consider two
bounds similar if one of the bounds contains some symbols and the other
doesn't contain any symbols. This is important to handle the situation
when unsigned symbolic value B (e.g., integer representing a boolean) is
pruned with zero (False). Before the prune, the value is represented
by an interval [B(u), B(u)] ("u" stands for unsigned) and is considered
precise. The result of the prune is [B(u), 0] and wasn't considered
precise before this change. Also, after the change, we only consider
two non-symbolic bounds similar if they are equal.

Note that [B(u), 0] is equal to [0, 0] (in the contexts where B(u) is
symbolic), but we need to keep the symbolic information in the interval
to be able to propagate (reachability) information from callee to caller.

Also note that the result of a prune of a signed interval [I(s), I(s)]
with [0, 0] is [max (I(s), 0), min (I(s), 0)] which is also equal to
[0, 0]. In that case, Bounds.are_similar returned True even before the
change.

(2) and (3) are important to make evaluation of expressions more precise.

* infer/src/bufferoverrun/bounds.ml
(are_similar): consider two bounds as similar also when one of the
bounds contains symbols and the other doesn't contain any symbol.
Consider two non-symbolic bounds as similar only if they are equal.

* infer/src/bufferoverrun/itv.ml
(is_false): instead of comparing the interval with False (zero) directly,
return True if the interval is <= and >= zero. As the result, is_false
returns True also for [B(u), 0] and [max (I(s), 0), min (I(s), 0)].
(eq_sem): Return True also in the case when the bounds of two intervals
itv1 and itv2 are not equal and it holds that itv1 <= itv2 <= itv1.
(ne_sem): Use eq_sem instead of duplicating the logic of eq_sem.