[I] U430-024 Track set of possible targets for unknown pointers.
When (un)known pointer is weakly updated with unknown, keep its
(un)known set of targets meaning that the pointer is unknown, but
it is more likely that it points to these targets than to other
targets.

This makes it possible to utilize the following heuristics - when
a deref of an unknown pointer is updated, these tracked targets are
updated as well (other possible, but not tracked targets are
unsoundly kept unupdated)

Example:

ptr = I'Access;
unknwn_call (ptr); -- may update ptr, ptr is unknown, but I is
                   -- still tracked as its alias
ptr'access = 4; -- we want to update I

Fixes diff on SA18-039 and on eurocontrol.